combine datatypes.py and endpoints.py into api.py
